web/IP-logger/requirements.txt
anyio==3.6.1
click==8.1.3
fastapi==0.85.0
h11==0.14.0
httptools==0.5.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.1
psycopg2-binary==2.9.3
pydantic==1.10.2
PyJWT==2.5.0
python-dotenv==0.21.0
python-multipart==0.0.5
PyYAML==6.0
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
typing_extensions==4.3.0
uvicorn==0.18.3
uvloop==0.17.0
watchfiles==0.17.0
websockets==10.3
